requirements.txt
Flask==2.0.2
urllib3==1.25.11
Flask_SQLAlchemy==2.5.1
Pillow==8.4.0
PyMySQL==1.0.2
requests==2.26.0
SQLAlchemy==1.3.24
